requirements.txt
streamlit_chat
boto3
openai
langchain
python-dotenv
gradio
git+https://github.com/openai/whisper.git
pymilvus